+++ lib/ext_video_player.dart
@@ -181,6 +181,7 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
       {this.package, this.closedCaptionFile, this.videoPlayerOptions})
       : dataSourceType = DataSourceType.asset,
         formatHint = null,
+        youtubeVideoQuality = null,
         super(VideoPlayerValue(duration: null));
 
   /// Constructs a [VideoPlayerController] playing a video from obtained from
@@ -191,7 +192,10 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
   /// **Android only**: The [formatHint] option allows the caller to override
   /// the video format detection code.
   VideoPlayerController.network(this.dataSource,
-      {this.formatHint, this.closedCaptionFile, this.videoPlayerOptions})
+      {this.formatHint,
+      this.closedCaptionFile,
+      this.videoPlayerOptions,
+      this.youtubeVideoQuality})
       : dataSourceType = DataSourceType.network,
         package = null,
         super(VideoPlayerValue(duration: null));
@@ -206,6 +210,7 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
         dataSourceType = DataSourceType.file,
         package = null,
         formatHint = null,
+        youtubeVideoQuality = null,
         super(VideoPlayerValue(duration: null));
 
   int _textureId;
@@ -213,6 +218,8 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
   /// The URI to the video file. This will be in different formats depending on
   /// the [DataSourceType] of the original video.
   final String dataSource;
+
+  final VideoQuality youtubeVideoQuality;
 
   /// **Android only**. Will override the platform's generic file format
   /// detection with whatever is set here.
@@ -253,7 +260,7 @@ class VideoPlayerController extends ValueNotifier<VideoPlayerValue> {
     _lifeCycleObserver.initialize();
     _creatingCompleter = Completer<void>();
 
-    VideoQuality quality = VideoQuality.medium360;
+    VideoQuality quality = youtubeVideoQuality ?? VideoQuality.medium360;
 
     String finalYoutubeUrl = dataSource;
     if (_getIdFromUrl(dataSource) != null) {